refactor: split selinux applets and table-drive mkfs aliases
Reorganize two applet areas without changing behavior, help text, or output.

internal/applets/securityutils/selinux: split the single large selinux.go by
concern. Move the shared backend (Backend interface, fsBackend, fixtureBackend,
Mode and its String method) into backend.go, the read-only query applets
(getenforce, selinuxenabled, sestatus, getsebool, matchpathcon) into query.go,
and the privileged/gated applets into privileged.go. selinux.go keeps the
package doc, Command type, name constants, constructors, Synopsis and Run
dispatch.

mke2fs and mkfs_vfat: drive the alias front-ends (mke2fs/mkfs.ext2 and
mkfs.vfat/mkdosfs) from a small name -> config metadata table, keeping the
shared filesystem-builder logic untouched. Synopsis now resolves through the
table for every alias.

Closes #761
Closes #779
