+++ internal/applets/util-linux/mke2fs/mke2fs_test.go
@@ -134,6 +134,12 @@ func TestAlias(t *testing.T) {
 	if New().Name() != "mke2fs" || NewMkfsExt2().Name() != "mkfs.ext2" {
 		t.Errorf("alias names wrong")
 	}
+	// Both aliases resolve their synopsis through the shared name -> config table.
+	for _, c := range []*Command{New(), NewMkfsExt2()} {
+		if c.Synopsis() != "Create an ext2 filesystem" {
+			t.Errorf("%s synopsis = %q", c.Name(), c.Synopsis())
+		}
+	}
 }
 
 func TestErrors(t *testing.T) {
